Bump to lemmy v4 beta 1

+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/BlockPerson.kt
@@ -33,7 +33,7 @@ data class BlockPerson (
 
     /* The person id. */
     @SerializedName("person_id")
-    val personId: kotlin.Double
+    val personId: kotlin.Long
 
 ) {
